The SonarQube rules that *Quality-time* uses to query SonarQube for t… (#1651)
The SonarQube rules that *Quality-time* uses to query SonarQube for the 'commented out code', 'complex units', 'long units', 'many parameters', and 'suppressed violations' metrics are no longer a parameter that the user can change. The reason is that it's hardly ever necessary to change these parameters and at the same time it's very easy to accidentally remove a rule and get incorrect results as a consequence. The used rules are documented in the metrics and sources overview (docs/METRICS_AND_SOURCES.md). Closes #1648.