The target column would show the technical debt target value instead … (#4173)
The target column would show the technical debt target value instead of the target value when technical debt was accepted. Fixed by always displaying the target value. In addition, if technical debt is accepted, the target value has a grey label. On hovering, a popup shows the technical debt target value and the technical debt end date, if any. Fixes #4171.