Modify Proxy-pod to run Nginx as non Root user #8356
The listening port of Nginx is now configurable using de environment variable NGINX_LISTEN. Default is still port 80, for backwards compatiblity. By pushing port to 8080, Nginx proces can be without root privileges. Using the envsubst command is a workaround because using variables in Nginx configuration is restricted. For security reasons it would be preferable to move to port 8080.